requirements.txt
PyYAML==6.0
Requests==2.31.0
wget==3.2
